seperate annopages into new component
